+++ src/lib/components/copy/Copier.tsx
@@ -1,14 +1,14 @@
 import type { LayoutProps } from "@chakra-ui/react";
 
 import { CustomIcon } from "../icon";
-import { AmpEvent, AmpTrack } from "lib/services/amplitude";
+import { AmpTrackCopier } from "lib/services/amplitude";
 
 import { CopyTemplate } from "./CopyTemplate";
 
 interface CopierProps {
   value: string;
   copyLabel?: string;
-  className?: string;
+  className: string;
   display?: LayoutProps["display"];
   ml?: string;
 }
@@ -29,7 +29,7 @@ export const Copier = ({
         display={display}
         cursor="pointer"
         marginLeft={ml}
-        onClick={() => AmpTrack(AmpEvent.USE_COPIER)}
+        onClick={() => AmpTrackCopier(className)}
         name="copy"
         boxSize="12px"
       />